requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
huggingface-hub==0.25.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
praat-parselmouth @ git+https://github.com/shoma-knot/Parselmouth.git@1a61e04c79febd28469cd6ce0805d1c02cdd3981
pycparser==2.22
pyparsing==3.1.4
PySoundFile==0.9.0.post1
python-dateutil==2.9.0.post0
pyworld @ git+https://github.com/shoma-knot/Python-Wrapper-for-World-Vocoder.git@8919feb53e9cac1d87aef921a8d25e9a60fe87ab
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
Resemblyzer==0.1.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
webrtcvad==2.0.10
